+++ CHANGELOG.md
@@ -197,3 +197,5 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
     - `im_runlen`
     - `im_cv2`
     - `im_alphablend`
+
+## Version 0.6.3 - Unreleased